tests/requirements.txt
ddt==1.1.2
nose==1.3.7
pytest==3.5.0
coverage==4.4.1
pytest-sugar==0.8.0
pytest-django==3.1.2
django_jenkins==0.110.0
